Updated migrations so they auto-increments

+++ src/data/migrations/20221112_createStockTable.js
@@ -20,7 +20,7 @@ SELECT * FROM WebServices_Taak.Stock;
 module.exports = {
     up: async (knex) => {
         await knex.schema.createTable(tables.stock, (table) => {
-            table.bigint('stockId').unsigned().notNullable();
+            table.increments('stockId').unsigned().notNullable();
             table.string('symbol', 5).notNullable();
             table.string('name', 60).notNullable();
             table.string('industry', 60).defaultTo(null);
